requirements.txt
django==1.6.10
wheel==0.22.0
django-simple-audit==0.1.14
South==0.8.2
dbaas_credentials==0.1.0
python-faasclient>=0.2.1
paramiko==1.13.0
